push

+++ backend/utils/removeimg.js
@@ -1,5 +1,11 @@
 import fs from "fs";
 
 export const removeimg = (path) => {
-  fs.unlinkSync(path);
+  try {
+    if (fs.existsSync(path)) {
+      fs.unlinkSync(path);
+    }
+  } catch (error) {
+    console.error("Error removing image:", error);
+  }
 };
